git-svn-id: http://morgoth.zemris.fer.hr/svn/vhdllab/vhdllab/trunk/vhdllab@293 cbae1e92-611c-0410-ab7b-b19ac7dee622

+++ src/web/onClient/hr/fer/zemris/vhdllab/applets/schema/SchemaSerializableInformation.java
@@ -22,7 +22,7 @@ public class SchemaSerializableInformation {
 	 */
 	public ArrayList<AbstractSchemaWire> wireList;
 	
-	public int componentNameCounter;
+	public long componentNameCounter;
 	public int wireNameCounter;
 }
